Internal: switch to single config.ts file
With better-docs integration, .ts files can be parsed by jsdoc, so separate
module declarations can be dropped
