+++ openpgp.d.ts
@@ -11,7 +11,7 @@
 
 import type { WebStream as GenericWebStream, NodeWebStream as GenericNodeWebStream } from '@openpgp/web-stream-tools';
 import enums from './src/enums.ts';
-import config, { type Config, type PartialConfig } from './src/config/index.ts';
+import config, { type Config, type PartialConfig } from './src/config.ts';
 
 export { enums, config, Config, PartialConfig };
