--- src/config/index.js
@@ -1,10 +1,0 @@
-/**
- * @fileoverview This object contains global configuration values.
- * @see module:config/config
- * @module config
- * @access public
- */
-
-import config from './config.ts';
-
-export default config;
