+++ src/packet/public_key.js
@@ -18,7 +18,7 @@
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 
 import KeyID from '../type/keyid.js';
-import defaultConfig from '../config/index.ts';
+import defaultConfig from '../config.ts';
 import { computeDigest, parsePublicKeyParams, serializeParams } from '../crypto/index.js';
 import enums from '../enums.ts';
 import util from '../util.js';
