+++ src/packet/sym_encrypted_integrity_protected_data.js
@@ -21,7 +21,7 @@ import { cipherMode, getRandomBytes, getCipherParams, computeDigest } from '../c
 import computeHKDF from '../crypto/hkdf.js';
 import enums from '../enums.ts';
 import util from '../util.js';
-import defaultConfig from '../config/index.ts';
+import defaultConfig from '../config.ts';
 
 import LiteralDataPacket from './literal_data.js';
 import CompressedDataPacket from './compressed_data.js';
