+++ src/packet/symmetrically_encrypted_data.js
@@ -21,7 +21,7 @@ import { readToEnd as streamReadToEnd, clone as streamClone } from '@openpgp/web
 import { cipherMode, getCipherParams } from '../crypto/index.js';
 import enums from '../enums.ts';
 import util from '../util.js';
-import defaultConfig from '../config/index.ts';
+import defaultConfig from '../config.ts';
 
 import LiteralDataPacket from './literal_data.js';
 import CompressedDataPacket from './compressed_data.js';
