+++ src/type/s2k/generic.js
@@ -16,7 +16,7 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 
-import defaultConfig from '../../config/index.ts';
+import defaultConfig from '../../config.ts';
 import { getRandomBytes, computeDigest } from '../../crypto/index.js';
 import enums from '../../enums.ts';
 import { UnsupportedError } from '../../packet/packet.js';
