+++ src/util.js
@@ -26,7 +26,7 @@
 import { concat as streamConcat, transform as streamTransform, concatUint8Array, isStream, isUint8Array } from '@openpgp/web-stream-tools';
 import { createRequire } from 'module'; // Must be stripped in browser built
 import enums from './enums.ts';
-import defaultConfig from './config/index.ts';
+import defaultConfig from './config.ts';
 
 const debugMode = (() => {
   try {
